✨ *refact(exceptions): reorganizar e documentar exceções do sistema*
Reorganiza as classes de exceção, movendo MailError para um novo módulo e removendo o módulo de exceções obsoleto. Adiciona documentação para a pasta de exceções.

Files Changed:
- *crawjud_app/addons/mail.py*: atualiza a importação de MailError para o novo caminho.
- *crawjud_app/common/exceptions/README.md*: cria documentação para a pasta de exceções.
- *crawjud_app/common/exceptions/__init__.py*: adiciona a classe BaseExceptionCeleryAppError.
- *crawjud_app/common/exceptions/mail.py*: cria a classe MailError com documentação.
- *crawjud_app/exceptions/__init__.py*: remove módulo de exceções obsoleto.
