--- crawjud_app/exceptions/__init__.py
@@ -1,7 +1,0 @@
-"""Módulo de tratamento de exceptions do crawjud_app."""
-
-from __future__ import annotations
-
-
-class BaseExceptionCeleryAppError(Exception):
-    """Base exception class for Celery app errors."""
